move ERB monkey patches to their own file
