+++ actionview/lib/action_view/template/handlers/erb.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require "strscan"
-require "active_support/core_ext/string"
+require "active_support/core_ext/erb/util"
 
 module ActionView
   class Template
